requirements.txt
configobj==5.0.6
fs==0.5.4
influxdb==2.12.0
pyserial==3.0.1
simplejson==3.8.1
val==0.6
zeroconf==0.17.5
